requirements.txt
aiohttp==3.6.2
aiopg==1.0.0
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.9
multidict==4.7.6
psycopg2-binary==2.8.5
PyYAML==5.3.1
SQLAlchemy==1.3.17
yarl==1.4.2
